.ssh/key_patterns: Moved from private repo
This is used by `bin/cjslogin`; the functionality is likely to be
replaced by `ckssh` or perhaps even just `.ssh/config`. We're keeping
it here for the moment so that a) this repo works standalone, since
`cjslogin` requires that the file be present and b) so we have an
example of the format in case we need it.