.fvwm/config: Remove xterm/rxvt -title/-n options
`-title` sets the title for only a fraction of a second until a shell
prompt appears, at which point that uses my xtitle program to change it.

The icon names are never used since I don't display iconized windows.

I do leave the -title for QuickShell because that's how the window
manager detects that it should take the focus. However, I also now
make sure the title is properly set after the shell prompt prints by
setting XTITLE=QuickShell in the environment.