.config/.../xfce4-panel.xml: Move to opt/ dir to get this out of the way
We use this when intially setting up a desktop system, but after that
it always seems to have local changes that are either not compatible
with other systems (e.g., width of 35 for 4K monitors versus 25 for
others) or unimportant changes, such as the list of visible items in
the status area that's updated every time a new program uses the
status area.

So we dump this off to the side where we can still use it as an
initial config on new desktop systems but otherwise ignore it.

We may later figure out a better way to deal with this.