.fvwm/config: Main menu xterm hotkey → T, rework quit/exit menu
We make the hotkey for xterm `t` (which is right beside the `r` for
urxvt), freeing up `x` to become our Exit menu and leaving `q`
unbound, which helps avoid accidently bringing up the quit menu when
we mean quickmove from the window menu instead.

The exit menu now has no letter keys bound except for `n` not to quit
and `r` to restart, both of which are relatively harmless. Again, this
helps avoid accidents.

The rarely used 'restart with xterm instead of fvwm' now has no
binding (we have to move down to it with arrow or `j`). The fvwm exit
and various session exit/save options are now on number keys where
they're reachable but well out of the way.