.Xresources Rxvt.font: Switch to Noto Mono as default
Noto Mono is slightly lighter weight than Source Code Pro; this looks
a bit better on 4K screens and considerably better on smaller screens.

This is still possibly an intermediate step; it's an improvement over
Source Code Pro but I still need to look at Inconsolata. That looks
definitely better on 4K screens, but not on lower resolution screens,
though some work (possibly including :scale=1.1 or similar) may fix
that. I did try Source Code Pro Light, too, but that was, though good
looking, too light to read comfortably in many situations.

We seem to need to explicitly set line space for Noto Mono; to get the
same spacing as Source Code Pro with lineSpace=0, on smaller monitors
we need to set lineSpace=1 and on 4K lineSpace=2.

This gives the usual vertically-maximized 106 line terminal on 4K
screens. On my 1440x900 ThinkPad X201s I move down to a 62-line
terminal (lineSpacing=0 or a standard xterm gives 67 lines), but we'll
see if I can live with that because it does look more comfortable.

The lineSpaceing setting is in pixels and thus must be an integer
value. (It appears to be rounded down.)