dh/update: Fetch and integrate changes from master branch
This is a `dh/update` script for branches that contain changes to be
added to the latest version of `master`. After a fetch it attempts to
automatically integrate any changes from master by rebasing the
current branch on top of it.

If the rebase fails for any reason, `git rebase --abort` will be
executed to return the branch to its original state, leaving things
running as they were. A manual rebase will need to be done to bring in
the master branch changes.

To install this, just cherry-pick this commit on to your branch.