.Xresources: Configure optional solarized colors
This is the code from https://github.com/altercation/solarized.git but
with #ifdefs to enable the parts that actually change the display. (Use
`xrdb -DSOLARIZED` etc. to enable them.)

1. SOLARIZED_DARK determines whether we define the colors for dark or
   light. However, this appears to be the wrong way to do it; see below.
2. SOLARIZED determines whether we set the colors (for all applications)
   or just use the defaults.

This method of defining the colors differently (e.g., making S_base03
a light color rather than a dark one) to switch to light mode is
consistent with neither the description on the web page nor with what
vim `:colorscheme solarized` expects, making `set background=light`
give a dark background.

While I like the idea of a more consistent color scheme, experience so
far with solarized has been that a) it has less contrast than I find
comfortable for reading and b) bad interactions with programs that
assume a more 'default' set of ANSI terminal colors, such as htop.