bin/xresetup: Overwrite rather than merge .Xresources
I am not clear on why I was using `xrdb -merge`, since this would
still overwrite changes with the original values from the .Xresources
where they existed, but would not delete additional resources not in
the .Xresources file. There isn't a way to say, 'keep changes set
outside .Xresources but incorporate changes in .Xresources to values
that are the ones it had previously set', so it's best just to reload
a full fresh set.

If this becomes bothersome we can perhaps add an option not to touch
the resources at all or allow us to pass extra arguments to `xrdb`.