.fvwm/config Window Ops menu: hjkl to move focus in that direction
h/j/k/l in the window menu now move the focus to the closest
non-overlapped window in the given direction, moving the mouse pointer
along with it (to the right side and half-way down the window). The
mouse pointer is important both because other commands such as layer
up/down apply to the window under the mouse pointer, to help give a
secondary visual indication of what window has the focus, and to avoid
confusion when doing a mouse move after the refocus.

This requires considerable movement of other keys, due to moving
functions previously on h/j/k/l and then the knock-on moves from that.
Most moves are to adjacent keys.
- Layer down/up keys, previously on j/k, move to u/i. u can be
  remembered as "up."
- Insert Fn key, previously i, moves to o. Parallel (a bit) to my vim
  insert-mode special insert sequence, Ctrl-O Ctrl-I.
- Lock screen, l, moves to ;. Also added to Root menu so that it works
  the same from either menu.
- Quick shell, previously ;, moves to '.
- Maximize both directions, previously h, moves to b.
- Destroy window, previously b, moves to x.

That last one is not the safest; perhaps we should move delete and
destroy window functions to a submenu.