.fvwm/config: Reddish title bar for active window
The previous dark title bar on for the active window was hard to find on a
4k screen because it was only slightly different from the background
colour, and also the contents were hard to read because of low contrast.

Instead, make it similar to the non-active title bars except with a more
reddish tint, making it stand out quite well as well as being easier to
read. It's still a fairly pale colour so it's not too jarring when placed
against the rest of the colour scheme.