bin/disp: x2 w/VGA→HDMI modes now working and cleaned up
The inability to do 2560x1440 turned out to be due to the frequency limits
of the adapter. The best we can do is 1920x1200, so we add a custom `x2h`
mode for that, though we still get a bit of hash on the display. The
standard 1920x1080 `x2f` mode still works fine and is clean. Other attempts
at higher resolution modes have been removed.