.fvwm/config: Title bar font Source Code Pro → Noto Serif
For some reason, after a Debian upgrade 11 → 12 on agnostic (X201s), the
title bar text stop displaying. Tweaking the font showed that this was
clearly a font issue, though I had Source Code Pro installed already into
~/.fonts/. Inconsolata and Noto Mono both look kinda crap, but Noto Serif
looks quite good, and is quite it's compact (because proportional). I'll do
it, because it almost reminds me of a NeXT or something.