bin/genmon-vclock: Add -h option for horizontal mode.
This is a lot wider than vertical mode is tall because I don't see any
obvious way to get the small "nn CT mm UT" parts to stack on top of
each other, though the font size is (probably just) small enough for
that.

Nonetheless, it's better than not having this clock at all now that
I'm trying horizontal mode on my 4K monitor because that looks as if
it works better for me with the panel on the right side of the screen.

And yes, the coding and error-handling here is a bit crap. If I do
much more work on this, that needs to be cleaned up. Though if I do
much more work on this, it may start to make sense, instead of just
using the XML-style markup, to write a real program to lay things out
properly (if that can be done).