.fvwm/config: Fix xfce4-panel startup to always work
Using --restart works when it's running, but if for some reason it's
not, --restart just produces an error dialogue and then runs until you
click "OK". So instead we check to see if it's running, start it with
no options if not, and use --restart only if it's already running.