bin/4up: Expand tabs
The program is not working very well, however, badly enough that it
might be considered just broken. What we may want to do really is
re-work this to have the "set up terminal, optionally log in to
another host, and either way execute some commands before going
interactive" thing in a separate script.