bin/tmjoin: Try (and fail) to close window w/o keypress if no error
As documented in more detail in the comments in the code, we see no way to
determine if there's "no error", so we still always wait for the user to
press Enter, though now with code that now demonstrates how we can check
the process exit code. (The current code we check for will never appear,
which is why it still always waits.)

Also improve the formatting of the code we run in the new windows.