dot/Xresources: Remove custom 16-color terminal colors
We leave them commented out for reference. This was mainly so that the
old Vim configuration had nicer colours in 16-color mode, but that
caused issues when pair-programming in shared terminals because my
color scheme could produce colors that were hard to read on in
terminals using the standard 16 colors. (That's solved by moving to a
256-color configuration in Vim, since virtually everyone's using
256-color terminals these days.)