.fvwm/config: Rework menu-windowops, esp. hjkl focus moves to top
I'd totally forgotten that I have these very handy focus move commands
and should use them more. Put them at the top so I am more likely to
see and remember them.

Now use tabs in the menu labels so that the columns line up more
nicely. Add a modeline to remind me of what settings I need to change
to avoid tabs for indents in the rest of the file. (This doesn't work
unless `modeline` is set--which it's not by default in Debian--or it's
manually executed.)